+++ src/index.ts
@@ -1,3 +1,4 @@
 export * as commands from "./commands"
 export * as printers from "./printers"
 export * as labels from "./labels"
+export { configureSVGRasterizer } from "./helpers/ImageProcessor"